remove errors and warnings
git-svn-id: https://svn.salilab.org/imp/trunk@12784 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ applications/integrative_docking/em2d_dock/em2d_single_score.cpp
@@ -88,7 +88,7 @@ aligned images (default = 0.4)");
   for(unsigned int i=0; i<image_files.size(); i++) {
     Image2D<> image(image_files[i]);
     image.get_largest_connected_component();
-    image.pad(image.get_width()*1.4, image.get_height()*1.4);
+    image.pad((int)(image.get_width()*1.4), (int)(image.get_height()*1.4));
     image.center();
     image.average();
     image.stddev();
